+++ tests/c/atom.c
@@ -14,7 +14,7 @@ int main() {
     CHFL_ATOM* a = chfl_atom("He");
     assert(a != NULL);
 
-    float mass=0, charge=0;
+    double mass=0, charge=0;
     assert(!chfl_atom_mass(a, &mass));
     assert(fabs(mass - 4.002602) < 1e-6);
     assert(!chfl_atom_charge(a, &charge));
